refactor(tests): consolidate turn-recovery tests; drop desync naming
